Added mcp, added to README.md.

+++ net/minecraft/client/gui/FontRenderer.java
@@ -371,7 +371,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string with a shadow.
      */
-    public int drawStringWithShadow(String text, float x, float y, int color)
+    public int drawStringWithShadow(String text, double x, double y, int color)
     {
         return this.drawString(text, x, y, color, true);
     }
@@ -387,7 +387,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, float x, float y, int color, boolean dropShadow)
+    public int drawString(String text, double x, double y, int color, boolean dropShadow)
     {
         this.enableAlpha();
 
@@ -639,7 +639,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Render single line string by setting GL color, current (posX,posY), and calling renderStringAtPos()
      */
-    private int renderString(String text, float x, float y, int color, boolean dropShadow)
+    private int renderString(String text, double x, double y, int color, boolean dropShadow)
     {
         if (text == null)
         {
@@ -667,8 +667,8 @@ public class FontRenderer implements IResourceManagerReloadListener
             this.green = (float)(color & 255) / 255.0F;
             this.alpha = (float)(color >> 24 & 255) / 255.0F;
             this.setColor(this.red, this.blue, this.green, this.alpha);
-            this.posX = x;
-            this.posY = y;
+            this.posX = (float) x;
+            this.posY = (float) y;
             this.renderStringAtPos(text, dropShadow);
             return (int)this.posX;
         }
